+++ cpp/core/add_observation_noise.cpp
@@ -1,12 +1,12 @@
 #include "add_observation_noise.h"
 #if 0
 //http://moby.ihme.washington.edu/bradbell/mat2cpp/randn.cpp.xml
-MatrixXf randn(int m, int n) 
+MatrixXd randn(int m, int n) 
 {	
 	// use formula 30.3 of Statistical Distributions (3rd ed)
 	// Merran Evans, Nicholas Hastings, and Brian Peacock
 	int urows = m * n + 1;
-	MatrixXf u(urows, 1);
+	MatrixXd u(urows, 1);
 
 	//u is a random matrix
 	for (int r=0; r<urows; r++) {
@@ -15,11 +15,11 @@ MatrixXf randn(int m, int n)
 		}
 	}
 	
-	MatrixXf x(m,n);
+	MatrixXd x(m,n);
 
 	int i, j, k;
-	float pi = 4. * std::atan(1.f);
-	float square, amp, angle;
+	double pi = 4. * std::atan(1.f);
+	double square, amp, angle;
 	k = 0;
 	for(i = 0; i < m; i++)
 	{	for(j = 0; j < n; j++)
@@ -39,38 +39,38 @@ MatrixXf randn(int m, int n)
 	return x;
 }
 
-MatrixXf rand(int m, int n) 
+MatrixXd rand(int m, int n) 
 {	
-	MatrixXf x(m,n);	
+	MatrixXd x(m,n);	
 	int i, j;
-	float rand_max = float(RAND_MAX);
+	double rand_max = double(RAND_MAX);
 
 	for(i = 0; i < m; i++) {	
 		for(j = 0; j < n; j++)
-			x(i, j) = float(std::rand()) / rand_max;
+			x(i, j) = double(std::rand()) / rand_max;
 	} 
 	return x;
 }
 #endif
 //add random measurement noise. We assume R is diagnoal matrix
-void add_observation_noise(vector<Vector2f> &z, Matrix2f R, int addnoise)
+void add_observation_noise(vector<Vector2d> &z, Matrix2d R, int addnoise)
 {
-    float LO = -1.0f;
-    float HI = 1.0f;
+    double LO = -1.0f;
+    double HI = 1.0f;
 
 	if (addnoise == 1){
 		int len = z.size();	
 		if (len > 0) {
-			//MatrixXf randM1 = nRandMat::randn(1,len);
-            MatrixXf randM1(1,len);
+			//MatrixXd randM1 = nRandMat::randn(1,len);
+            MatrixXd randM1(1,len);
             for (int i=0; i< len; i++) {
-                float r3 = LO + (float)rand()/((float)RAND_MAX/(HI-LO));
+                double r3 = LO + (double)rand()/((double)RAND_MAX/(HI-LO));
                 randM1(0,i) = r3; 
             }
-			//MatrixXf randM2 = nRandMat::randn(1,len);
-			MatrixXf randM2(1,len);
+			//MatrixXd randM2 = nRandMat::randn(1,len);
+			MatrixXd randM2(1,len);
             for (int j=0; j< len; j++) {
-                float r4 = LO + (float)rand()/((float)RAND_MAX/(HI-LO));
+                double r4 = LO + (double)rand()/((double)RAND_MAX/(HI-LO));
                 randM2(0,j) = r4; 
             }
             //cout<<"randM1"<<endl;
