+++ cpp/core/data_associate_known.cpp
@@ -2,8 +2,8 @@
 #include <iostream>
 
 //z is range and bearing of visible landmarks
-void data_associate_known(vector<Vector2f> z, vector<int> idz, VectorXf &table, int Nf, \
-		vector<Vector2f> &zf, vector<int> &idf, vector<Vector2f> &zn) 
+void data_associate_known(vector<Vector2d> z, vector<int> idz, VectorXd &table, int Nf, \
+		vector<Vector2d> &zf, vector<int> &idf, vector<Vector2d> &zn) 
 {
 	idf.clear();
 	vector<int> idn;
@@ -12,7 +12,7 @@ void data_associate_known(vector<Vector2f> z, vector<int> idz, VectorXf &table, 
 
 	for (i =0; i< idz.size(); i++){
 		ii = idz[i];
-		VectorXf z_i;
+		VectorXd z_i;
 		if (table(ii) ==-1) { //new feature
 			z_i = z[i];
 			zn.push_back(z_i);
